Segregated buy token logic into dedicated hook
