+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/TradeTokenForm/types.ts
@@ -1,6 +1,7 @@
-import { TokenView } from '@hicommonwealth/schemas';
+import { ExtendedCommunity, TokenView } from '@hicommonwealth/schemas';
 import { ChainBase } from '@hicommonwealth/shared';
 import { SupportedCurrencies } from 'helpers/currency';
+import NodeInfo from 'models/NodeInfo';
 import { z } from 'zod';
 import useTradeTokenForm from './useTradeTokenForm';
 
@@ -8,8 +9,6 @@ export enum TradingMode {
   Buy = 'buy',
   Sell = 'sell',
 }
-
-export type TradeTokenFormProps = ReturnType<typeof useTradeTokenForm>;
 
 const TokenWithCommunity = TokenView.extend({
   community_id: z.string(),
@@ -29,6 +28,15 @@ export type UseTradeTokenFormProps = {
   addressType?: ChainBase;
   onTradeComplete?: () => void;
 };
+
+export type UseBuyTradeProps = UseTradeTokenFormProps & {
+  chainNode: NodeInfo;
+  tokenCommunity?: z.infer<typeof ExtendedCommunity>;
+  selectedAddress?: string;
+  commonFeePercentage: number;
+};
+
+export type TradeTokenFormProps = ReturnType<typeof useTradeTokenForm>;
 
 export type AddressBalanceProps = Pick<
   ReturnType<typeof useTradeTokenForm>,
