+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/TradeTokenForm/useTradeTokenForm.ts
@@ -1,14 +1,10 @@
+import { ExtendedCommunity } from '@hicommonwealth/schemas';
 import { commonProtocol } from '@hicommonwealth/shared';
 import { notifyError, notifySuccess } from 'controllers/app/notifications';
 import useRunOnceOnCondition from 'hooks/useRunOnceOnCondition';
 import NodeInfo from 'models/NodeInfo';
 import { useMemo, useState } from 'react';
 import { useGetCommunityByIdQuery } from 'state/api/communities';
-import {
-  useFetchTokenUsdRateQuery,
-  useGetUserEthBalanceQuery,
-} from 'state/api/communityStake';
-import { useBuyTokenMutation } from 'state/api/launchPad';
 import useSellTokenMutation from 'state/api/launchPad/sellToken';
 import { fetchCachedNodes } from 'state/api/nodes';
 import {
@@ -16,9 +12,9 @@ import {
   useGetERC20BalanceQuery,
 } from 'state/api/tokens';
 import useUserStore from 'state/ui/user';
-import useJoinCommunity from 'views/components/SublayoutHeader/useJoinCommunity';
-import './TradeTokenForm.scss';
+import { z } from 'zod';
 import { TradingMode, UseTradeTokenFormProps } from './types';
+import useBuyTrade from './useBuyTrade';
 
 const COMMON_PLATFORM_FEE_PERCENTAGE = 5; // make configurable when needed
 
@@ -27,7 +23,6 @@ const useTradeTokenForm = ({
   addressType,
   onTradeComplete,
 }: UseTradeTokenFormProps) => {
-  const [baseCurrencyBuyAmount, setBaseCurrencyBuyAmount] = useState<number>(0);
   const [tradingMode, setTradingMode] = useState<TradingMode>(
     tradeConfig.mode || TradingMode.Buy,
   );
@@ -51,20 +46,6 @@ const useTradeTokenForm = ({
     (n) => n.ethChainId === commonProtocol.ValidChains.SepoliaBase,
   ) as NodeInfo; // this is expected to exist
 
-  const { linkSpecificAddressToSpecificCommunity } = useJoinCommunity();
-
-  const { data: ethToCurrencyRateData, isLoading: isLoadingETHToCurrencyRate } =
-    useFetchTokenUsdRateQuery({
-      tokenSymbol: 'ETH',
-    });
-  const ethToCurrencyRate = parseFloat(
-    ethToCurrencyRateData?.data?.data?.amount || '0.00',
-  );
-
-  const ethBuyAmount = baseCurrencyBuyAmount / ethToCurrencyRate;
-  const commonPlatformFeeForBuyTradeInEth =
-    (COMMON_PLATFORM_FEE_PERCENTAGE / 100) * ethBuyAmount;
-
   const { data: tokenCommunity, isLoading: isLoadingTokenCommunity } =
     useGetCommunityByIdQuery({
       id: tradeConfig.token.community_id,
@@ -72,43 +53,13 @@ const useTradeTokenForm = ({
       includeNodeInfo: true,
     });
 
-  // imp: this query uses CommunityStakes helper to get eth price, but its
-  // a generic query so no need to initiate a separate Launhpad helper
-  const {
-    data: selectedAddressEthBalance = `0.0`,
-    isLoading: isLoadingUserEthBalance,
-  } = useGetUserEthBalanceQuery({
-    chainRpc: tokenCommunity?.ChainNode?.url || '',
-    ethChainId: tokenCommunity?.ChainNode?.eth_chain_id || 0,
-    walletAddress: selectedAddress || '',
-    apiEnabled: !!(selectedAddress && tokenCommunity),
-  });
-
   useRunOnceOnCondition({
     callback: () => setSelectedAddress(userAddresses[0]),
     shouldRun: userAddresses.length > 0 && !selectedAddress,
   });
 
-  const { mutateAsync: buyToken, isLoading: isBuyingToken } =
-    useBuyTokenMutation();
-
   const { mutateAsync: createTokenTrade, isLoading: isCreatingTokenTrade } =
     useCreateTokenTradeMutation();
-
-  const onBaseCurrencyBuyAmountChange = (
-    change: React.ChangeEvent<HTMLInputElement> | number,
-  ) => {
-    if (typeof change == 'number') {
-      setBaseCurrencyBuyAmount(change);
-    } else {
-      const value = change.target.value;
-
-      if (value === '') setBaseCurrencyBuyAmount(0);
-      // verify only numbers with decimal (optional) are present
-      else if (/^\d+(\.\d+)?$/.test(value))
-        setBaseCurrencyBuyAmount(parseFloat(value));
-    }
-  };
 
   const onTradingModeChange = (mode: TradingMode) => {
     setTradingMode(mode);
@@ -118,59 +69,19 @@ const useTradeTokenForm = ({
     setSelectedAddress(address);
   };
 
-  const handleTokenBuy = async () => {
-    try {
-      // this condition wouldn't be called, but adding to avoid typescript issues
-      if (
-        !baseNode?.url ||
-        !baseNode?.ethChainId ||
-        !selectedAddress ||
-        !tokenCommunity
-      ) {
-        return;
-      }
-
-      // buy token on chain
-      const payload = {
-        chainRpc: baseNode.url,
-        ethChainId: baseNode.ethChainId,
-        amountEth: ethBuyAmount * 1e18, // amount in wei
-        walletAddress: selectedAddress,
-        tokenAddress: tradeConfig.token.token_address,
-      };
-      const txReceipt = await buyToken(payload);
-
-      // create token trade on db
-      await createTokenTrade({
-        eth_chain_id: baseNode?.ethChainId,
-        transaction_hash: txReceipt.transactionHash,
-      });
-
-      // join user's selected address to community
-      const isMemberOfCommunity = user.addresses.find(
-        (x) => x.community.id === tokenCommunity.id,
-      );
-      if (!isMemberOfCommunity) {
-        await linkSpecificAddressToSpecificCommunity({
-          address: selectedAddress,
-          community: {
-            base: tokenCommunity.base,
-            iconUrl: tokenCommunity.icon_url || '',
-            id: tokenCommunity.id,
-            name: tokenCommunity.name,
-          },
-        });
-      }
-
-      // update user about success
-      notifySuccess('Transactions successful!');
-
-      onTradeComplete?.();
-    } catch (e) {
-      notifyError('Failed to buy token');
-      console.log('Failed to buy token => ', e);
-    }
-  };
+  const {
+    amounts: buyTradeAmounts,
+    handleTokenBuy,
+    isBuyActionPending,
+    selectedAddressEthBalance,
+  } = useBuyTrade({
+    chainNode: baseNode,
+    selectedAddress,
+    commonFeePercentage: COMMON_PLATFORM_FEE_PERCENTAGE,
+    tradeConfig,
+    onTradeComplete,
+    tokenCommunity: tokenCommunity as z.infer<typeof ExtendedCommunity>,
+  });
 
   // sell mode logic start --- {
   const [tokenSellAmount, setTokenSellAmount] = useState<number>(0); // can be fractional
@@ -242,11 +153,9 @@ const useTradeTokenForm = ({
   // flag to indicate if something is ongoing
   const isActionPending =
     isLoadingTokenCommunity ||
-    isLoadingUserEthBalance ||
-    isBuyingToken ||
+    isBuyActionPending ||
     isLoadingUserTokenBalance ||
     isSellingToken ||
-    isLoadingETHToCurrencyRate ||
     isCreatingTokenTrade;
 
   const onCTAClick = () => {
@@ -270,27 +179,7 @@ const useTradeTokenForm = ({
     // functionality is done in most "onChange" handlers above
     trading: {
       amounts: {
-        buy: {
-          invest: {
-            baseCurrency: {
-              name: tradeConfig.currency, // USD/GBP etc
-              amount: baseCurrencyBuyAmount,
-              onAmountChange: onBaseCurrencyBuyAmountChange,
-              presetAmounts: tradeConfig.presetAmounts,
-              unitEthExchangeRate: ethToCurrencyRate,
-              toEth: ethBuyAmount,
-            },
-            insufficientFunds:
-              ethBuyAmount > parseFloat(selectedAddressEthBalance),
-            commonPlatformFee: {
-              percentage: `${COMMON_PLATFORM_FEE_PERCENTAGE}%`,
-              eth: commonPlatformFeeForBuyTradeInEth,
-            },
-          },
-          gain: {
-            token: 100, // TODO: hardcoded for now - blocked token pricing
-          },
-        },
+        buy: buyTradeAmounts,
         sell: {
           invest: {
             // not to be confused with "Base" network on ethereum
@@ -323,8 +212,8 @@ const useTradeTokenForm = ({
         value: selectedAddress,
         balances: {
           eth: {
-            value: selectedAddressEthBalance,
-            isLoading: isLoadingUserEthBalance,
+            value: selectedAddressEthBalance.value,
+            isLoading: selectedAddressEthBalance.isLoading,
           },
           selectedToken: {
             value: selectedAddressTokenBalance,
